rvbbit/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rvbbit"
version = "2.0.0"
description = "RVBBIT: Declarative agent framework for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm",
    "chdb",
    "pypdf",
    "duckdb",  # Still used by smart_sql_run tool for agent queries
    "pandas",
    "pyarrow",
    "python-dotenv",
    "pydantic",
    "rich",
    "jinja2",
    "altair",
    "vl-convert-python",
    "plotly",
    "kaleido",
    "miniaudio",
    "pillow",
    "gunicorn[gevent]",
    "jsonschema",
    "readchar",  # For Hot or Not interactive rating UI
    "miniaudio",  # For ElevenLabs TTS audio playback
    "requests",  # For ElevenLabs API calls and Rabbitize REST API
    "ruamel.yaml",  # For YAML cascade/tool definition support
]

[project.optional-dependencies]
# Extra dependencies for specific features
browser = [
    # Rabbitize is npm-based, so we can't install it here
    # But we list it for documentation purposes
    # Install with: npm install -g rabbitize
]
docker = [
    "docker",  # For linux_shell and run_code tools
]
all = [
    "docker",
]

[project.scripts]
rvbbit = "rvbbit.cli:main"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:litellm.*",
]
markers = [
    "requires_llm: marks tests that require LLM API access (OpenRouter)",
    "requires_clickhouse: marks tests that require ClickHouse database",
    "integration: marks integration tests that need external services",
]
# Default: run all tests except those requiring external services
# Use -m "not integration" to skip integration tests
# Use -m "" to run all tests
